UPDATE: Window and Feature. Both classes now have unmnaged and managed impl. Those should be extended when developing new extensions. UPDATE: SpecRollOff and Hamming. Conforming new base class impl. UPDATE: Single or float is now the supported type. REMOVE: Sound. Completely outdated.

+++ src/Extensions/Windows/Hamming/Hamming.cs
@@ -52,10 +52,10 @@ namespace MusiC.Extensions.Windows
 		{
 		}
 		
-		override public double Factory(int n)
+		override public Single Factory(int n)
 		{
 			//return 0.53836 - 0.46164 * (Math.Cos(2 * Math.PI * n / (WindowSize - 1)));
-			return 0.54 - 0.46 * (Math.Cos(2 * Math.PI * n / (WindowSize - 1)));
+			return (Single) (0.54f - 0.46f * (Math.Cos(2.0f * Math.PI * n / (WindowSize - 1))));
 		}
 	}
 }
